+++ Plutarch/Internal/Term.hs
@@ -21,7 +21,6 @@ module Plutarch.Internal.Term (
   pplaceholder,
   punsafeCoerce,
   punsafeBuiltin,
-  punsafeConstant,
   punsafeConstantInternal,
   compile,
   compileWithInternalConfig,
@@ -661,10 +660,6 @@ punsafeCoerce (Term x) = Term x
 
 punsafeBuiltin :: UPLC.DefaultFun -> Term s a
 punsafeBuiltin f = Term \_ -> pure $ mkTermRes $ RBuiltin f
-
-{-# DEPRECATED punsafeConstant "Use `pconstant` instead." #-}
-punsafeConstant :: Some (ValueOf PLC.DefaultUni) -> Term s a
-punsafeConstant = punsafeConstantInternal
 
 punsafeConstantInternal :: Some (ValueOf PLC.DefaultUni) -> Term s a
 punsafeConstantInternal c = Term \_ ->